fix factories for docs generation

+++ database/migrations/2023_12_08_235304_create_accounts_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
             $table->unsignedBigInteger('currency_id')->nullable();
             $table->string('name');
             $table->string('bank');
-            $table->integer('balance');
+            $table->bigInteger('balance');
             $table->timestamps();
             $table->softDeletes();
